arm: dts: aspeed: Pull BMC ready pin
Author: John Wang

The FAN board for Barreleye G2 has a circuit to force all FANs
speed to full if BMC FW is not ready or watchdog timeout.

The state of BMC is decided by GPIO BMC_READY_N.
Level is HIGH by default.
The state of watchdog timeout reset is decided by WDTRST1.
Level is LOW by default.

We need pull BMC_READY_N to low so that fan speed can be
controlled by sysfs.

Change-Id: I3ac32902f68c2dd114dee75089eef162c6426ae0
Signed-off-by: John Wang <hsienchiang@gmail.com>
Signed-off-by: Doyle Huang <doyle.sy.huang@mail.foxconn.com>